extract errors
