+++ public/components/ajax-step-flow/ajax-step-flow__slide.js
@@ -2,6 +2,12 @@
 
 import { showErrorText } from 'components/form/form-feedback-wrap';
 import { getUrlErrors } from 'js/get-url-errors';
+import {
+  MalformedResponseError,
+  MalformedHtmlError,
+  BackendError,
+  ContextError
+} from 'js/errors';
 import { getRaven } from 'components/sentry/sentry';
 import {
   formRoutes as validAjaxFormRoutes,
@@ -23,48 +29,6 @@ type ParsedResponse = {
   },
   url: string
 };
-
-class BackendError extends Error {
-  errors: string[];
-  constructor(errors) {
-    super('ERR_BACKEND_ERROR');
-    // $FlowFixMe
-    this.constructor = BackendError;
-    // $FlowFixMe
-    this.__proto__ = BackendError.prototype; // eslint-disable-line no-proto
-    this.errors = errors;
-  }
-}
-
-class ContextError extends Error {
-  context: {};
-}
-
-class MalformedResponseError extends ContextError {
-  constructor(request) {
-    super('ERR_MALFORMED_RESPONSE');
-    // $FlowFixMe
-    this.constructor = MalformedResponseError;
-    // $FlowFixMe
-    this.__proto__ = MalformedResponseError.prototype; // eslint-disable-line no-proto
-    this.context = {
-      request
-    };
-  }
-}
-
-class MalformedHtmlError extends ContextError {
-  constructor(html) {
-    super('ERR_MALFORMED_HTML');
-    // $FlowFixMe
-    this.constructor = MalformedHtmlError;
-    // $FlowFixMe
-    this.__proto__ = MalformedHtmlError.prototype; // eslint-disable-line no-proto
-    this.context = {
-      html
-    };
-  }
-}
 
 const getSlide = ($wrapper: HTMLElement): HTMLElement => {
   const $slide = $wrapper.querySelector(selector);
